api/requirements.txt
-i https://pypi.org/simple
certifi==2022.9.14; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
colorama==0.4.5; platform_system == 'Windows'
flask==2.2.2; python_version >= '3.7'
flask-cors==3.0.10
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
werkzeug==2.2.2; python_version >= '3.7'
